+++ lib/data/comment/CommentList.class.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace data\comment;
+
+use data\DatabaseObjectList;
+
+class CommentList extends DatabaseObjectList {
+}